# improve rewriteAttrsAction speed [AS-803]