--- a/core/src/main/scala/org/broadinstitute/dsde/rawls/dataaccess/slick/AttributeComponent.scala
+++ b/core/src/main/scala/org/broadinstitute/dsde/rawls/dataaccess/slick/AttributeComponent.scala
@@ -355,6 +355,36 @@ trait AttributeComponent {
       val toSaveAttrMap = toPrimaryKeyMap(attributesToSave)
       val existingAttrMap = toPrimaryKeyMap(existingAttributes)
 
+      // note that currently-existing attributes will have a populated id e.g. "1234", but to-save will have an id of "0"
+      // therefore, we use this ComparableRecord class which omits the id when checking equality between existing and to-save.
+      // note this does not include transactionId for AttributeScratchRecords. We do not expect AttributeScratchRecords
+      // here, and transactionId will eventually be going away, so don't bother
+      object ComparableRecord {
+        def fromRecord[OWNER_ID](rec: RECORD): ComparableRecord[OWNER_ID] = {
+          // TODO: davidan: is the asInstanceOf safe here?? I'm not sure why rec.ownerId doesn't compile on its own
+          new ComparableRecord[OWNER_ID](rec.ownerId.asInstanceOf[OWNER_ID], rec.namespace, rec.name, rec.valueString, rec.valueNumber, rec.valueBoolean,
+            rec.valueJson, rec.valueEntityRef, rec.listIndex, rec.listLength, rec.deleted, rec.deletedDate)
+        }
+      }
+
+      case class ComparableRecord[OWNER_ID] (
+        ownerId: OWNER_ID,
+        namespace: String,
+        name: String,
+        valueString: Option[String],
+        valueNumber: Option[Double],
+        valueBoolean: Option[Boolean],
+        valueJson: Option[String],
+        valueEntityRef: Option[Long],
+        listIndex: Option[Int],
+        listLength: Option[Int],
+        deleted: Boolean,
+        deletedDate: Option[Timestamp]
+      )
+
+      // create a set of ComparableRecords representing the existing attributes
+      val existingAttributesSet: Set[ComparableRecord[OWNER_ID]] = existingAttributes.toSet.map(r => ComparableRecord.fromRecord[OWNER_ID](r))
+
       val existingKeys = existingAttrMap.keySet
 
       // insert attributes which are in save but not exists
@@ -363,32 +393,10 @@ trait AttributeComponent {
       // delete attributes which are in exists but not save
       val attributesToDelete = existingAttrMap.filterKeys(! toSaveAttrMap.keySet.contains(_))
 
-      // update attributes which are in both to-save and currently-exists, but have different values.
-      // note that currently-existing attributes will have a populated id e.g. "1234", but to-save will have an id of "0"
-      // therefore, use a comparison function that looks at everything except id
-      // note this does not compare transactionId for AttributeScratchRecords. We do not expect AttributeScratchRecords
-      // here, and transactionId will eventually be going away, so don't bother
-      // TODO: should this compare function move closer to the AttributeRecord class?
-      def equalRecords(left: RECORD, right: RECORD): Boolean = {
-        // compare everything except id
-        left.name == right.name &&
-        left.valueString == right.valueString &&
-        left.valueNumber == right.valueNumber &&
-        left.valueBoolean == right.valueBoolean &&
-        left.valueJson == right.valueJson &&
-        left.valueEntityRef == right.valueEntityRef &&
-        left.listIndex == right.listIndex &&
-        left.listLength == right.listLength &&
-        left.namespace == right.namespace &&
-        left.ownerId == right.ownerId &&
-        left.deleted == right.deleted &&
-        left.deletedDate == right.deletedDate
-      }
-
       val attributesToUpdate = toSaveAttrMap.filter {
         case (k, v) =>
             existingKeys.contains(k) && // if the attribute doesn't already exist, don't attempt to update it
-            !existingAttributes.exists(equalRecords(_, v)) // if the attribute exists and is unchanged, don't update it
+            !existingAttributesSet.contains(ComparableRecord.fromRecord[OWNER_ID](v)) // if the attribute exists and is unchanged, don't update it
       }
 
       // collect the parent objects (e.g. entity, workspace) that have writes, so we know which object rows to re-calculate