--- a/core/src/test/resources/fixtures/batchUpsert/README.md
+++ b/core/src/test/resources/fixtures/batchUpsert/README.md
@@ -1,0 +1,2 @@
+* mediumtext-initial.json: anonymized data with 2304 rows and
+  59 columns. Some of the columns contain values 200-300 characters long.