@davidangb 2021-06-19
merging with one thumb instead of two as an override; will get a second review soon after this is released as a hotfix.